+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 3.4.10
+
+### Path Changes
+-   Updated UI Libraries
+-   Fixed `thunderstorm` weather icons
+-   Improved UI `Popout` component
+-   Created `PeriodSelector` UI component
+
 ## 3.4.9
 
 ### Path Changes